Задание 10 спринта. Доработка согласно замечаниям
