+++ .idea/misc.xml
@@ -12,7 +12,7 @@
         <entry key="..\:/PlaylistMaker1/app/src/main/res/layout/activity_media.xml" value="0.3625" />
         <entry key="..\:/PlaylistMaker1/app/src/main/res/layout/activity_search.xml" value="0.33" />
         <entry key="..\:/PlaylistMaker1/app/src/main/res/layout/activity_settings.xml" value="0.3625" />
-        <entry key="..\:/PlaylistMaker1/app/src/main/res/layout/track_layout.xml" value="0.33" />
+        <entry key="..\:/PlaylistMaker1/app/src/main/res/layout/track_layout.xml" value="0.5" />
         <entry key="..\:/PlaylistMaker1/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.2435" />
       </map>
     </option>
