+++ app/src/main/java/com/example/playlistmaker/MusicAdapter.kt
@@ -1,0 +1,18 @@
+package com.example.playlistmaker
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+
+class MusicAdapter(private val music: List<Track>) : RecyclerView.Adapter<MusicViewHolder>() {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MusicViewHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.track_layout, parent, false)
+        return MusicViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: MusicViewHolder, position: Int) {
+        holder.bind(music[position])
+    }
+
+    override fun getItemCount() = music.size
+}
