+++ app/src/main/java/com/example/playlistmaker/MusicViewHolder.kt
@@ -1,0 +1,36 @@
+package com.example.playlistmaker
+
+import android.view.View
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.bumptech.glide.load.resource.bitmap.RoundedCorners
+
+class MusicViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+
+    private val trackName: TextView
+    private val artistName: TextView
+    private val trackTime: TextView
+    private val artworkUrl100: ImageView
+
+    init {
+        trackName = itemView.findViewById(R.id.trackName)
+        artistName = itemView.findViewById(R.id.artistName)
+        trackTime = itemView.findViewById(R.id.trackTime)
+        artworkUrl100 = itemView.findViewById(R.id.artworkUrl100)
+    }
+
+    fun bind(model: Track) {
+
+        trackName.text = model.trackName
+        artistName.text = model.artistName
+        trackTime.text = model.trackTime
+        // Работа с библиотекой
+        Glide.with(itemView).load(model.artworkUrl100).centerCrop()
+            .placeholder(R.drawable.ic_noconnection).transform(RoundedCorners(15))
+            .into(artworkUrl100)
+
+    }
+
+}
