+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -1,25 +1,10 @@
 package com.example.playlistmaker
 
-data class Track(private val trackName: String,
-            private val artistName: String,
-            private val trackTime: String,
-            private val artworkUrl100: String){
-
-
-    fun getTrackName(): String {
-        return trackName
-    }
-
-    fun getArtistName(): String {
-        return artistName
-    }
-
-    fun getTrackTime(): String {
-        return trackTime
-    }
-
-    fun getArtworkUrl100(): String {
-        return artworkUrl100
-    }
+data class Track(
+    val trackName: String,
+    val artistName: String,
+    val trackTime: String,
+    val artworkUrl100: String
+) {
 
 }
