+++ app/src/main/res/drawable/ic_ellipse_1.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="3dp"
+    android:height="3dp"
+    android:viewportWidth="3"
+    android:viewportHeight="3">
+  <path
+      android:pathData="M1.5,1.5m-1.5,0a1.5,1.5 0,1 1,3 0a1.5,1.5 0,1 1,-3 0"
+      android:fillColor="#AEAFB4"/>
+</vector>
