+++ app/src/main/res/layout/track_layout.xml
@@ -16,70 +16,50 @@
             android:id="@+id/artworkUrl100"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:src="@drawable/ic_baseline_music_off_24"
-            android:contentDescription="TODO" />
+            android:layout_gravity="center"
+            android:src="@drawable/ic_noconnection" />
 
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
             android:orientation="vertical">
 
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/trackName" />
-
-                <TextView
-                    android:id="@+id/trackName"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:gravity="end"
-                    android:text="@string/NoInternet" />
-
-            </LinearLayout>
+            <TextView
+                android:id="@+id/trackName"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:fontFamily="@font/ys_display_bold"
+                android:textSize="16sp" />
 
             <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/ArtistName" />
+                android:layout_width="wrap_content"
+                android:layout_height="15dp"
+                android:orientation="horizontal">
 
                 <TextView
                     android:id="@+id/artistName"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
-                    android:text="@string/NoInternet" />
+                    android:fontFamily="@font/ys_display_light"
+                    android:textSize="11sp" />
 
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/duration" />
+                <ImageView
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:src="@drawable/ic_ellipse_1"
+                    android:layout_marginLeft="5dp"
+                    android:layout_marginRight="5dp"/>
 
                 <TextView
                     android:id="@+id/trackTime"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
-                    android:text="@string/NoInternet" />
+                    android:fontFamily="@font/ys_display_light"
+                    android:textSize="11sp" />
 
             </LinearLayout>
 
-
         </LinearLayout>
-
 
     </LinearLayout>
